refactor parsing functions
