+++ helpers/calculate.js
@@ -1,5 +1,5 @@
 import Handlebars from 'handlebars'
-import { hoursFromNow } from '../lib/parse'
+import { hoursFromNow } from '../lib/parse/time'
 
 
 Handlebars.registerHelper('dueStatus', function(dueDate) {
