+++ lib/parse/promise.js
@@ -1,4 +1,5 @@
-import { parseDate, parseText } from '../parse'
+import { parseDate, parseTimezone } from './time'
+import parseText from './text'
 
 // FIXME allow passing multiple overriding params
 
@@ -25,4 +26,15 @@ export function parsePromise({ urtext, timezone }) {
     console.log('parsePromise', urtext, promise)
     return promise
   }
+}
+
+export default function parsePromiseWithIp({ urtext, ip }) {
+  return new Promise((resolve, reject) => {
+    parseTimezone(ip).then((timezone) => {
+      resolve(parsePromise({ urtext, timezone }))
+    })
+    .catch (() => {
+      reject('failed to parse timezone from IP') // new Error('text')
+    })
+  })
 }
